Fergan

+++ src/NameServer/Resolver.java
@@ -16,7 +16,7 @@ public class Resolver implements ResolverInterface
 
 
 	@Override
-	public String getIP(int nodeId) throws RemoteException, InvalidParameterException
+	public String getIP(short nodeId) throws RemoteException, InvalidParameterException
 	{
 		if(NameServer.getInstance().map.containsKey(nodeId))
 		{
@@ -29,7 +29,7 @@ public class Resolver implements ResolverInterface
 	}
 
 	@Deprecated
-	public void removeFromTree(int nodeId) throws RemoteException
+	public void removeFromTree(short nodeId) throws RemoteException
 	{
 		if(NameServer.getInstance().map.containsKey(nodeId))
 		{
